Check if required binaries exist in PATH

Some of cragr's functionalities require certain third-party binaries to exist in PATH. For example, selectively loading cfDNA fragments from a certain region under the hood invokes tabix to complete the task. This function checks whether certain binaries exist.

check_binaries(binaries = "tabix", verbose = FALSE, stop_on_fail = FALSE)

Arguments

binaries
Character vector indicating what binaries to check. Default is tabix.

verbose
Logical value indicating whether to output detailed messages. Default is FALSE.

stop_on_fail
Stop the script completely if the result is FALSE.

Value

Logical value. TRUE if all required binaries can be found in PATH.