Randomly place intervals of fixed width across genome

This function is aware of chromosome size differences and intervals are randomly generated in proportion to chromosome sizes. In another word, larger chromosomes will have more randomly generated intervals.

make_random_bed(
  n,
  interval_width,
  chrom_sizes = c("hg19", "hg38"),
  chrom = NULL,
  seed = NULL,
  sort = TRUE
)

Arguments

n
Number of intervals to generate.

interval_width
An integer value of the width of the interval.

chrom_sizes
Provide the chromosome sizes data. If hg19 or hg38, the embedded chromosome sizes for hg19 or hg38 will be used. Otherwise, it need to be the path of a user-provided file.

chrom
A character vector. If provided, only generate random intervals for the specified chromosome(s).

seed
An integer seed for the random number generator. If NULL, the seed is randomly chosen.

sort
Logical value. Should the random BED data be sorted?

Value

A data.table of generated intervals.

References

Manual page for bedtools random: https://bedtools.readthedocs.io/en/latest/content/tools/random.html

Examples

# Generate 100 500kbp-intervals for chr20 and chr22, hg38
result <- make_random_bed(100, 500e3L, chrom_sizes = "hg38", chrom = c("chr20", "chr22"))
head(result)

#>    chrom    start      end
#> 1: chr20   250273   750273
#> 2: chr20  1851209  2351209
#> 3: chr20  2736167  3236167
#> 4: chr20  5175998  5675998
#> 5: chr20  9298753  9798753
#> 6: chr20 10628140 11128140

# Generate 100 500bp unsorted intervals for all chromosomes of hg19 (default reference genome)
result <- make_random_bed(100, interval_width = 500, chrom_sizes = "hg19", sort = FALSE)
head(result)

#>    chrom     start       end
#> 1:    11  45998718  45999218
#> 2:    11  60996043  60996543
#> 3:    11  68842273  68842773
#> 4:     5 146164531 146165031
#> 5:     5  69965452  69965952
#> 6:     5  35340911  35341411