Make windows along genome

Make windows with the size of window_size along genome. Must provide chrom_sizes so that intervals are restricted by chromosome sizes.

make_windows(window_size, chrom_sizes = c("hg19", "hg38"), chrom = NULL)

Arguments

window_size
An integer value. Size of the windows in base pairs.

chrom_sizes
Provide the chromosome sizes data. If hg19 or hg38, the embedded chromosome sizes for hg19 or hg38 will be used. Otherwise, it need to be the path of a user-provided file.

chrom
A character vector. If provided, only generate windows for the specified chromosome(s).

Value

A data.table of generated intervals.

Examples

# Generate 500kbp intervals for chr20 and chr22, hg38
result <- make_windows(window_size = 500e3L, chrom_sizes = "hg38", chrom = c("chr20", "chr22"))
head(result)

#>    chrom   start     end
#> 1: chr20       0  500000
#> 2: chr20  500000 1000000
#> 3: chr20 1000000 1500000
#> 4: chr20 1500000 2000000
#> 5: chr20 2000000 2500000
#> 6: chr20 2500000 3000000

# Generate 5Mbp intervals for all chromosomes of hg19 (default reference genome)
result <- make_windows(window_size = 5e6L)
head(result)

#>    chrom    start      end
#> 1:     1        0  5000000
#> 2:     1  5000000 10000000
#> 3:     1 10000000 15000000
#> 4:     1 15000000 20000000
#> 5:     1 20000000 25000000
#> 6:     1 25000000 30000000