Load a BED-format file

This function loads the input file as a data.table object. The file can be either local or remote, and can be either plain text or gzip-compressed. Furthermore, this function supports range-loading by providing a genomic range in the following syntax: "chr1:1-100".

read_bed(file_path, range = NULL, ...)

Arguments

file_path
Path to the data file. It can be either a local file, or a remote URL.

range
A genomic range character vector. Must follow standard genomic range notation format, e.g. chr1:1001-2000

...
Other arguments to be passed to data.table::fread().

Details

Note: for loading remote data files, currently this function depends on tabix.c 0.2.5, which doesn't not support HTTPS protocol. In the next step, I plan to turn to htslib, and the this function can load remote data files through HTTPS.

See also

Examples

bedtbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))
head(bedtbl)

#>    chrom start end score
#> 1:    21     3   6     8
#> 2:    21    10  16     5
#> 3:    21    12  17     4
#> 4:    21    15  22     5
#> 5:    21    22  25     7
#> 6:    21    26  30     7

bedtbl <- read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
                   range = "1:3001-4000")
head(bedtbl)

#>    chrom start  end score1 score2
#> 1:     1  2925 3011    106    181
#> 2:     1  3003 3092     88    193
#> 3:     1  3091 3193    118    212
#> 4:     1  3164 3248     94    211
#> 5:     1  3232 3345    107    205
#> 6:     1  3300 3395     88    193

# Does not work. Currently HTTPS is not supported
# bedtbl <- read_bed("https://yourdomain.com/example-02.bedGraph.gz", 
#                    range = "1:1000-2000")