Slop intervals

Similar to bedtools slop, this function increases the size of each feature in a feature file by a user-defined number of bases. Will restrict the resizing to the size of the chromosome.

slop_bed(
  x,
  chrom_sizes = c("hg19", "hg38"),
  slop = 1L,
  slop_type = c("bp", "frac"),
  mode = c("both", "left", "right")
)

Arguments

x
A data.table object.

chrom_sizes
Provide the chromosome sizes data. If hg19 or hg38, the embedded chromosome sizes for hg19 or hg38 will be used. Otherwise, it need to be the path of a user-provided file. If NULL, the resizing will not be restricted to the size of the chromosome.

slop
The number of base pairs to slop, or the percentage of the interval's width (refer to bedtools slop's -pct argument).

slop_type
If bp, slop is an integer indicating the number of base pairs. If frac, slop is a float point number indicating the percentage.

mode
If both, the resizing takes place on both ends of the interval. Otherwise, the resizing is only for the left end or the right end.

Value

A slopped data.tabe object.

References

Manual page of bedtools slop: https://bedtools.readthedocs.io/en/latest/content/tools/slop.html

Examples

# Load data
tbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))

# Slop by 10 on both ends
result <- slop_bed(tbl, slop = 10L, slop_type = "bp", mode = "both")

# Slop by 10 only on the right end
result <- slop_bed(tbl, slop = 10L, slop_type = "bp", mode = "right")

# Slop by 20% on the left end
result <- slop_bed(tbl, slop = 0.2, slop_type = "frac", mode = "left")