Load a BED-format file

This function loads the input file as a data.table object. The file can be either local or remote, and can be either plain text or gzip-compressed. Furthermore, this function supports range-loading by providing a genomic range in the following syntax: "chr1:1-100".

read_bed(
  file_path,
  range = NULL,
  compression = c("detect", "bgzip", "text", "other"),
  tabix_index = NULL,
  download_index = FALSE,
  genome = NULL,
  use_gr = TRUE,
  sep = "\t",
  ...
)

Arguments

file_path
Path to the data file. It can be either a local file, or a remote URL.

range
A genomic range character vector. Must follow standard genomic range notation format, e.g. chr1:1001-2000

compression
Indicate the compression type. If detect, this function will try to guess from file_path.

tabix_index
A character value indicating the location of the tabix index file. Can be either local or remote. If NULL, it will be derived from file_path.

download_index
Whether to download (cache) the tabix index at current directory.

genome
Specify the reference genome for the BED file. genome can be a valid genome name in GenomeInfoDb::Seqinfo(), e.g. GRCh37, or hs37-1kg, which is a genome shipped with this package, or any custom chromosome size files (local or remote). Here is a good resource for such files: https://github.com/igvteam/igv/tree/master/genomes/sizes.

use_gr
If TRUE, will read the data as a GenomicRanges object, otherwise a data.table object. Generally, we recommend using GenomicRanges.

sep
The separator between columns. By default, BED files are tab-delimited, and sep should be \t. However, sometimes you will encounter non-standard table files. In such cases, you need to specify the separator. If auto, read_bed will try to guess the separator. For more details, refer to data.table::fread().

...
Other arguments to be passed to data.table::fread().

Details

Note: for loading remote data files, currently this function depends on tabix.c 0.2.5, which doesn't not support HTTPS protocol. In the next step, I plan to turn to htslib, and the this function can load remote data files through HTTPS.

See also

Examples

bedtbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))
head(bedtbl)

#> GRanges object with 6 ranges and 1 metadata column:
#>       seqnames    ranges strand |     score
#>          <Rle> <IRanges>  <Rle> | <integer>
#>   [1]       21       4-6      * |         8
#>   [2]       21     11-16      * |         5
#>   [3]       21     13-17      * |         4
#>   [4]       21     16-22      * |         5
#>   [5]       21     23-25      * |         7
#>   [6]       21     27-30      * |         7
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

# Basic usage
bedtbl <- read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
                  range = "1:3001-4000")
head(bedtbl)

#> GRanges object with 6 ranges and 2 metadata columns:
#>       seqnames    ranges strand |    score1    score2
#>          <Rle> <IRanges>  <Rle> | <integer> <integer>
#>   [1]        1 2926-3011      * |       106       181
#>   [2]        1 3004-3092      * |        88       193
#>   [3]        1 3092-3193      * |       118       212
#>   [4]        1 3165-3248      * |        94       211
#>   [5]        1 3233-3345      * |       107       205
#>   [6]        1 3301-3395      * |        88       193
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Specify the reference genome
head(read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
              range = "1:3001-4000",
              genome = "hs37-1kg"))

#> GRanges object with 6 ranges and 2 metadata columns:
#>       seqnames    ranges strand |    score1    score2
#>          <Rle> <IRanges>  <Rle> | <integer> <integer>
#>   [1]        1 2926-3011      * |       106       181
#>   [2]        1 3004-3092      * |        88       193
#>   [3]        1 3092-3193      * |       118       212
#>   [4]        1 3165-3248      * |        94       211
#>   [5]        1 3233-3345      * |       107       205
#>   [6]        1 3301-3395      * |        88       193
#>   -------
#>   seqinfo: 84 sequences from hs37-1kg genome

head(read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
              range = "1:3001-4000",
              genome = "GRCh37"))

#> GRanges object with 6 ranges and 2 metadata columns:
#>       seqnames    ranges strand |    score1    score2
#>          <Rle> <IRanges>  <Rle> | <integer> <integer>
#>   [1]        1 2926-3011      * |       106       181
#>   [2]        1 3004-3092      * |        88       193
#>   [3]        1 3092-3193      * |       118       212
#>   [4]        1 3165-3248      * |        94       211
#>   [5]        1 3233-3345      * |       107       205
#>   [6]        1 3301-3395      * |        88       193
#>   -------
#>   seqinfo: 92 sequences from GRCh37 genome

head(read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
              range = "1:3001-4000",
              genome = "https://raw.githubusercontent.com/igvteam/igv/master/genomes/sizes/1kg_v37.chrom.sizes"))

#> GRanges object with 6 ranges and 2 metadata columns:
#>       seqnames    ranges strand |    score1    score2
#>          <Rle> <IRanges>  <Rle> | <integer> <integer>
#>   [1]        1 2926-3011      * |       106       181
#>   [2]        1 3004-3092      * |        88       193
#>   [3]        1 3092-3193      * |       118       212
#>   [4]        1 3165-3248      * |        94       211
#>   [5]        1 3233-3345      * |       107       205
#>   [6]        1 3301-3395      * |        88       193
#>   -------
#>   seqinfo: 86 sequences from 1kg_v37 genome

# Load remote BGZIP files with tabix index specified
head(read_bed("https://git.io/JYATB", range = "22:20000001-30000001", tabix_index = "https://git.io/JYAkT"))

#> GRanges object with 6 ranges and 2 metadata columns:
#>       seqnames            ranges strand |      score comp_method
#>          <Rle>         <IRanges>  <Rle> |  <numeric> <character>
#>   [1]       22 20000001-20500000      * | -0.0613535    cofrag/3
#>   [2]       22 20500001-21000000      * | -0.0422752    cofrag/3
#>   [3]       22 21000001-21500000      * |  0.0268175    cofrag/3
#>   [4]       22 21500001-22000000      * |  0.0386308    cofrag/3
#>   [5]       22 22000001-22500000      * |  0.0265141    cofrag/3
#>   [6]       22 22500001-23000000      * | -0.0425812    cofrag/3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths